+++ src/main.cpp
@@ -2,6 +2,7 @@
 #include "core/Game.hpp"
 #include "pong/PongGame.hpp"
 #include "SpaceInvaders/SpaceInvaders.hpp"
+#include "core/rasterizer/RasterizerDemo.hpp"
 
 void PlayPong(){
     PongGame pongGame(FVector2{800,450});
@@ -12,6 +13,11 @@ void PlaySpaceInvaders(){
     SpaceInvaders spaceInvaders;
     spaceInvaders.InitGame();
     spaceInvaders.Run();
+}
+void PlayRasterizerDemo(){
+    RasterizerDemo demo;
+    demo.InitGame();
+    demo.Run();
 }
 
 int main(){
@@ -46,6 +52,7 @@ int main(){
         DrawText("Try demo games [PRESS KEY]:", paddingLeft, 150, 20, WHITE);
         DrawText("[P] Pong", paddingLeft, 190, 20, WHITE);
         DrawText("[S] Space Invaders", paddingLeft, 230, 20, WHITE);
+        DrawText("[R] Rasterizer Demo", paddingLeft, 270, 20, WHITE);
 
         EndDrawing();
 
@@ -58,6 +65,12 @@ int main(){
         if (IsKeyPressed(KEY_S)) {
             CloseWindow();
             PlaySpaceInvaders();
+            break;
+        }
+
+        if (IsKeyPressed(KEY_R)) {
+            CloseWindow();
+            PlayRasterizerDemo();
             break;
         }
     }